Protected routes

+++ backend/controllers/tasks_controllers.js
@@ -3,7 +3,7 @@ const Task = require('../models/tasks_model')
 
 const getTasks = asyncHandler(async(req, res) => {
 
-    const tasks = await Task.find()
+    const tasks = await Task.find({ user: req.user._id })
 
     res.status(200).json({tasks})
 })
@@ -14,7 +14,8 @@ const createTasks = asyncHandler(async(req, res) => {
         throw new Error ('Please type a task description')
     }
     const task = await Task.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user._id
     })
     res.status(201).json({task})
 })
@@ -25,8 +26,13 @@ const updateTasks = asyncHandler(async(req, res) => {
         res.status(400)
         throw new Error('task does not exist')
     }
-    const taskUpdated = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true })
-    res.status(200).json(taskUpdated)
+    if (task.user.toString() !== req.user._id.toString()) {
+        res.status(401)
+        throw new Error('Unauthorized access')
+    } else {
+        const taskUpdated = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        res.status(200).json(taskUpdated)
+    }
 })
 
 const deleteTasks = asyncHandler(async(req, res) => {
@@ -35,8 +41,13 @@ const deleteTasks = asyncHandler(async(req, res) => {
         res.status(400)
         throw new Error('task does not exist')
     }
-    task.deleteOne()
-    res.status(200).json({id: req.params.id})
+    if (task.user.toString() !== req.user._id.toString()) {
+        res.status(401)
+        throw new Error('Unauthorized access')
+    } else {
+        task.deleteOne()
+        res.status(200).json({id: req.params.id})
+    }
 })
 
 module.exports = {
